Move Subtree tile to SUBTREELoader file
